fix(expo-router): unblock react server components from the package entry point
