+++ packages/expo-router/src/__rsc_tests__/index.test.tsx
@@ -1,0 +1,7 @@
+import { Link } from '../index';
+
+it('default import supports RSC', async () => {
+  const jsx = <Link href="/" />;
+
+  await expect(jsx).toMatchFlightSnapshot();
+});
