+++ packages/expo-router/src/react-navigation/bottom-tabs/TransitionConfigs/TransitionSpecs.tsx
@@ -1,12 +1,17 @@
-import { Easing } from 'react-native';
+import { Easing, type EasingFunction } from 'react-native';
 
 import type { TransitionSpec } from '../types';
+
+const lazyMemo = (factory: () => EasingFunction): EasingFunction => {
+  let cached: EasingFunction | undefined;
+  return (t: number) => (cached ??= factory())(t);
+};
 
 export const FadeSpec: TransitionSpec = {
   animation: 'timing',
   config: {
     duration: 150,
-    easing: Easing.in(Easing.linear),
+    easing: lazyMemo(() => Easing.in(Easing.linear)),
   },
 };
 
@@ -14,6 +19,6 @@ export const ShiftSpec: TransitionSpec = {
   animation: 'timing',
   config: {
     duration: 150,
-    easing: Easing.inOut(Easing.ease),
+    easing: lazyMemo(() => Easing.inOut(Easing.ease)),
   },
 };
